requirements.txt
python-dotenv==1.0.0
pydantic
fastapi
sqlalchemy
alembic
uvicorn[standard]
flake8
